requirements.txt
click
numpy
pyyaml
tqdm
scipy
torch==2.0.1